requirements.txt
# requirements.txt
aliyun-log-python-sdk
langfuse
python-dotenv